+++ source/CoordinateConversion/ArcMapAddinCoordinateConversion/ArcMapAddinCoordinateConversion.csproj
@@ -112,6 +112,7 @@
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
     <Compile Include="ValueConverters\IPointToStringConverter.cs" />
+    <Compile Include="ValueConverters\OppositeBoolValueConverter.cs" />
     <Compile Include="ViewModels\CollectTabViewModel.cs" />
     <Compile Include="ViewModels\ConvertTabViewModel.cs" />
     <Compile Include="ViewModels\TabBaseViewModel.cs" />
