+++ source/CoordinateConversion/ArcMapAddinCoordinateConversion/MainViewModel.cs
@@ -64,27 +64,4 @@ namespace ArcMapAddinCoordinateConversion.ViewModels
             }
         }
     }
-
-    //TODO move this to ValueConverters folder/namespace etc
-    public class BoolToOppositeBoolConverter : IValueConverter
-    {
-        #region IValueConverter Members
-
-        public object Convert(object value, Type targetType, object parameter,
-            System.Globalization.CultureInfo culture)
-        {
-            if (targetType != typeof(bool))
-                throw new InvalidOperationException("The target must be a boolean");
-
-            return !(bool)value;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter,
-            System.Globalization.CultureInfo culture)
-        {
-            throw new NotSupportedException();
-        }
-
-        #endregion
-    }
 }
