+++ source/CoordinateConversion/ArcMapAddinCoordinateConversion/ValueConverters/OppositeBoolValueConverter.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Windows.Data;
+
+namespace ArcMapAddinCoordinateConversion.ValueConverters
+{
+    public class BoolToOppositeBoolConverter : IValueConverter
+    {
+        #region IValueConverter Members
+
+        public object Convert(object value, Type targetType, object parameter,
+            System.Globalization.CultureInfo culture)
+        {
+            if (targetType != typeof(bool))
+                throw new InvalidOperationException("The target must be a boolean");
+
+            return !(bool)value;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter,
+            System.Globalization.CultureInfo culture)
+        {
+            throw new NotSupportedException();
+        }
+
+        #endregion
+    }
+}
